Fix install script to check existing links
Don't `rm tmuxinator/tmuxinator`!

+++ install.sh
@@ -4,11 +4,13 @@ for file in *
 do
     if [[ "${IGNORED_FILES[@]}" =~ "${file} " || "${IGNORED_FILES[${#IGNORED_FILES[@]}-1]}" == "${file}" ]]; then
         # whatever you want to do when arr contains value
-        echo "Should not link $file"
+        echo "Ignoring $file..."
     else
         link_from=$(pwd)/$file
         link_to=~/.$file
-        echo "linking $link_from to $link_to"
-        ln -s $link_from $link_to
+        if [ ! -L $link_to ]; then
+            echo "Linking $link_from to $link_to..."
+            ln -s $link_from $link_to
+        fi
     fi
 done
